seperation of logics
