+++ resources/views/layouts/app.blade.php
@@ -22,7 +22,6 @@
     <script src="https://code.jquery.com/jquery-3.5.1.min.js"
         integrity="sha384-ZvpUoO/+PpLXR1lu4jmpXWu80pZlYUAfxl5NsBMWOEPSjUn/6Z/hRTt8+pR6L4N2" crossorigin="anonymous">
     </script>
-    @vite(['resources/js/chessboard-1.0.0.min.js', 'resources/js/chess.ts', 'resources/js/app.js'])
 </body>
 
 </html>
