+++ vite.config.js
@@ -10,6 +10,7 @@ export default defineConfig({
                 "resources/css/chessboard-1.0.0.min.css",
                 "resources/js/chessboard-1.0.0.min.js",
                 "resources/js/chess.ts",
+                "resources/js/quick-solve.js",
             ],
             refresh: true,
         }),
